Add dashboard route
