+++ src/Dashboard.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Dashboard = () => {
+  return <div>Dashboard</div>;
+};
+
+export default Dashboard;
